+++ CHANGES.rst
@@ -21,6 +21,12 @@ API Changes
     different depending on the exact version of Python you are using.
     [#3308]
 
+- ``astropy.wcs``
+
+  - WCS allows slices of the form slice(None, x, y), which previously resulted
+    in an unsliced copy being returned (note: this was previously incorrectly
+    reported as fixed in v0.4.3) [#2909]
+
 
 0.4.3 (2015-01-15)
 ------------------
@@ -140,9 +146,6 @@ Bug Fixes
   - Fixed an issue with reconnecting to a SAMP Hub. [#2674]
 
 - ``astropy.wcs``
-
-  - WCS allows slices of the form slice(None, x, y), which previously resulted
-    in an unsliced copy being returned [#2909]
 
   - Invalid or out of range values passed to ``wcs_world2pix`` will
     now be correctly identified and returned as ``nan``